Rename plural to singular in api
