+++ client/src/components/recipes/CreateRecipeForm.svelte
@@ -1,6 +1,6 @@
 <script>
   import { onMount } from 'svelte';
-  import { categoriesApi, ingredientsApi, unitsOfMeasurementApi } from '@/api/index';
+  import { ingredientApi, unitOfMeasurementApi } from '@/api/index';
   import CreateStepsList from './CreateStepsList.svelte';
 
   let recipeForm = {
@@ -20,8 +20,8 @@
     // TODO: currently, if fetch fails, categories are set to [], but in the future this should inform user about error!
     // TODO: currently, if fetch fails, ingredients and units are set to [], but in the future this should inform user about error! (same story as categories)
     categories = await categoriesApi.fetchAll().catch(() => []);
-    ingredients = await ingredientsApi.getIngredients().catch(() => []);
-    unitsOfMeasurement = await unitsOfMeasurementApi.getUnitsOfMeasurement().catch(() => []);
+    ingredients = await ingredientApi.getIngredients().catch(() => []);
+    unitsOfMeasurement = await unitOfMeasurementApi.getUnitsOfMeasurement().catch(() => []);
   });
 </script>
